profile

+++ routes/index.js
@@ -100,6 +100,24 @@ router
         fs.unlink(files.fileRef.path)
         return res.redirect('/profile')
       }
+      if(fields.oldPassword.name || !fields.newPassword.name || !fields.confirmPassword.name) {
+        return res.redirect('/profile')
+      }
+      if(!fields.oldPassword.name || fields.newPassword.name || !fields.confirmPassword.name) {
+        return res.redirect('/profile')
+      }
+      if(!fields.oldPassword.name || !fields.newPassword.name || fields.confirmPassword.name) {
+        return res.redirect('/profile')
+      }
+      if(!fields.oldPassword.name || fields.newPassword.name || fields.confirmPassword.name) {
+        return res.redirect('/profile')
+      }
+      if(fields.oldPassword.name || fields.newPassword.name || !fields.confirmPassword.name) {
+        return res.redirect('/profile')
+      }
+      if(fields.oldPassword.name || !fields.newPassword.name || fields.confirmPassword.name) {
+        return res.redirect('/profile')
+      }
       fileName = path.join(upload, files.fileRef.name)
       fs.rename(files.fileRef.path, fileName, (err) => {
         if (err) {
